Move enrollment in the `EnrollmentController`
